src/test/java/it/eparlato/tictactoe/GameTest.java
package it.eparlato.tictactoe;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.assertThat;
import static org.mockito.Mockito.*;

class GameTest {
    private final Board board = mock(Board.class);
    private final Referee referee = mock(Referee.class);
    private Game game;
    public static final FieldCoordinates FIELD_COORDINATES = new FieldCoordinates(0, 0);

    @BeforeEach
    void setUp() {
        game = new Game(board, referee);
    }

    @Test
    void starts_with_a_new_game_state() {
        assertThat(game.state()).isEqualTo(GameState.NEW);
    }

    @Test
    void orchestrates_take_field_flow() {
        RefereeEvaluation refereeEvaluation = mock(RefereeEvaluation.class);
        when(referee.evaluation()).thenReturn(refereeEvaluation);

        game.takeField(FIELD_COORDINATES);

        verify(referee).check(board);
        verify(refereeEvaluation).applyOn(game);
    }

}